Fixing issue #14. The only method missing the link to the documentation was cp_apr.m
Also resolving linking issue in MATLAB with the documentation browser. If the complex linking command is spread across several lines, it works fine in the command window but does not work in the help browser. For example, `doc cp_apr` displays the command line help in the help brower, but the link to the documentation page would not work if the command had '...' continuation line commands in it.

Removed dead help page link for tt_gcp_fg_setup.m.

Fixed the order so that the doc page link comes after references and before see also.